backend/scripts/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
